disable NERD Tree plugin because it breaks in Vim 7.4
It shows an error message whenever I try to open gVim to a file, such as by double-clicking a file in Windows Explorer.

I hope NERD Tree updates soon and I enable it again. Or, if the problem is a conflict with my config, I hope that someone writes a guide for fixing that.