add eol, precedes, and conceal symbols to 'listchars'
I want eol so I can easily verify that apparently-blank lines are totally blank. And I still want trail so that I know I won't miss that a whitespace-only line has whitespace because the eol symbol is too far to the right. Though ideally, eol would show only on blank lines.

I rarely need precedes, but it should be useful when it shows.

I don't use conceal, but this character should be a good one if 'conceallevel' ever does happen to be set to 1.