Add TODO comment to look at two vimrcs
vim-sensible is actually a plugin, but I don’t want to use it because I want to change some of its settings, like `listchars`, and it doesn’t officially support being run early and overridden. So I’ll just copy the parts I like from it, unless it changes that policy.