Enable solarized color scheme to try it out
What I like about Solarized is the way it is simple but well-defined, it has been tested on many configurations, it is well-supported, and it is pretty. What I don’t like is that the background is not a neutral gray, but a shade of blue or yellow. I stuck with wombat so far because I was worried that I would eventually get sick of the blue hue from seeing it too much, whereas one can never have too much gray. But I want to try solarized again and see how it goes.

Remember that I can use the mapping `cob` from vim-sensible to change 'background' between "dark" and "light".

I also explicitly set the solarized_visibility of listchars to "normal", after trying out "low". "low" is meant to be used with 'cursorline' on, because the cursor line make the low-visibility characters readable, but I don’t have it on.