Reword the commented list of eol characters
The character I am using now is different from the other possibilities in that it actually fits within one monospace character cell.

If I later disable the eol character and put ¬ back in the list, I will reword the message again. See previous wordings in history for help with that.