Use :set’s multiple arguments to shorten autocmds
I am still keeping the RainbowParentheses commands on separate lines, since their are easier to read when aligned – I can ignore the common prefix. But these commands are better put on the same line, so it is clear that the settings are related, and probably to improve speed a tiny bit.