Add a <Leader> mapping to edit my vimrc
I used to use the GUI menu item, but now I think that is too hard compare to a mapping. On Windows, the menu item has a keyboard shortcut, so it’s easy enough there, but not on Mac. If I remember correctly, the menu command sequence is “Alt, E, V” on Windows, so I made the mapping use the same letters.