Record link to FastFold plugin for if I need it
I don’t need that plugin right now, so I won’t install it so that my folding is always correct. But if I edit large files with folding again, I will try installing it. If it works, I might keep it installed, but somehow configure it to only activate for large files which would otherwise be slowed by folding.