Replace Tabular plugin with EasyAlign
It seems easier to remember how to use it – I can type `=` instead of `/=`. And it supports a mapping, too.

I am using `+` as the visual-mode mapping, because it shares a key with `=`, which also aligns, and because that movement is still available with `<Enter>`. I didn’t define a normal-mode mapping because I don’t align enough for it to be worth overwriting the current definition of `ga`, even though I haven’t needed to use it yet.