Replace BufExplorer with Unite
I don’t use any of the advanced features of BufExplorer – not that there are many. I like Unite because it can display the buffer list in the current window, and supports fuzzy-search. It also shows NoName buffers by default, which means I don’t need that option for BufExplorer. And I am thinking of replacing Ctrl-P with Unite, so this is a step towards that.

The mapping was copied from http://www.codeography.com/2013/06/17/replacing-all-the-things-with-unite-vim.html. I used `<Leader>be` as the mapping, the same as BufExplorer, for familiarity. I tried `<Leader>b` first, but it turns out that both the bufkill plugin and one of my vimrc mappings already use that prefix.