Replace Gundo plugin with Undotree
I want Undotree for its feature to display and jump to points in history when the file was saved. I also like that it doesn’t require Python, so it will work on Windows.

I don’t like losing the option to automatically show the diff of the item under the cursor – it is slow, but sometimes useful – but I don’t think it’s worth keeping Gundo around just for that. And Undotree takes ~1.9ms instead of Gundo’s ~0.3ms to load at Vim startup, but I think it’s worth it.