`:set nocursorline` in NERDTree windows
The cursorline is misleading when enabled, because it highlights the file that was under the cursor when NERDTree was last in focus, even if the window has a totally different file open now. I expect highlighting to mean that the highlighted file is the open one, like in Sublime Text’s file sidebar.